backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pyodbc==5.0.1
pandas==2.1.1
python-multipart==0.0.6
pydantic==2.4.2
